Implemented lock character feature
It was necessary to add yet another atom to do it, but we made it! Now
clicking on a character will lock the definition on that character, and
hovering other characters will not update it until another click is
made. Furthermore, the locked character will show with a colored
background (the color is work in progress, for now it's orange).